fix: update errors
