+++ src/components/NotFound.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const NotFound = () => {
+  return <div>NotFound Page</div>;
+};
